requirements.txt
streamlit
PyMuPDF
PyPDF2
pikepdf
Pillow